modified: 중복제거/에러메시지 수정 - 중복제거 list comprehension --> for loop 로 변경 - Model.DoesNotExist 에러 외 발견한 에러 없음 --> Error Exception as e   제거
