+++ user/urls.py
@@ -1,8 +1,9 @@
 from django.urls import path
 
-from user.views import SignUpView, SignInView
+from user.views import SignUpView, SignInView, FollowUser
 
 urlpatterns = [
     path('/signup', SignUpView.as_view()),
     path('/signin', SignInView.as_view()),
+    path('/follow', FollowUser.as_view()),
 ]
